Added a weaponsmith and made some minor fixes
A weaponsmith now exists in the game.  It collects iron from storage
facilities and converts it into weapons.  Popups have also been fixed up
so that it is easier for new buildings to have popups added and the iron
mine popup was fixed to display the right text.  Farmable land has also
started getting implemented.