Updated health system and mud bath
The health system is nearly to where it needs to be.  It currently has some temp values so that orcs get sick more often than intended (for testing).  The mud bath has also been updated, but it still has a bug dealing with progress occuring when workers are not there.  The new implementation involves workers constantly spending time to keep the mud from drying up.  The building will help keep orcs healthy.