Updated how building houses works, added an immigrant queue, fixed up activate/deactivate on boat buildings
To build houses, the user can click to place one building or click and drag to place multiple buildings inside of a boxed area defined by the drag.  Added an immigrant queue so there isn't a stack of orc immigrants all created at the same instant.  There is a short time delay between when immigrants are allowed to be spawned.  Updated the Fishing Wharf and Boatyard to better take activation/deactivation into account.