Updated road placement and updated the warehouse art to visually fill
There is a new and improved way to build roads.  It replaces the old way.  Players can now drag while building a road to build a longer road.  If the player drags in both x and y directions, it will build as an 'L' shape.  The warehouse art has been updated so that the boxes better fit the art style of the rest of the game.  The warehouse will also visually fill and empty as it adds and removes goods.