Updated aqueduct placement and fixed issue with aqueducts + reservoirs emptying when they lose their water connection
The user can now place aqueducts in a straight line or 'L' shape by dragging the mouse while the left mouse button is held down.  Fixed issues with valid placement of aqueducts.  Fixed an issue where reservoirs and aqueducts might not empty out when the connection to a water source is removed.