Added game speeds, pause menu, and worked on World Map + Trade UI
The game can now be played at 1x, 1.5x, 2x speed.  The game can also be paused.  A new menu has been created which will pause the game and will eventually give the user access to changing settings and exiting the game.  Work has been continued on the UI for the world map and trading with other cities.