Started implementing Traders and fixed issues with Time in the game
Traders will now spawn in the city after opening a trade route, but logic for visiting a trading post and trading with the city still needs to be added.  Characters, Buildings, and camera Controls were updating off of game Time incorrectly.  The issue with game Time should be fixed now.